Add expression type property; fix error handling in add-expression POST API
Added the expression type functionality, which changed the database layout from
3 tables, one for each type, to one table 'expressions' which contains 'type'
column. Then changed the error handling of the API for POST method when adding
new expression so that no useless exceptions are thrown (exception is only when
the error is unexpected - submitting duplicate is very much expected). Also fixed
the throwing of Hibernate exception when submitting duplicate.