Refactoring of the word 'expression' to 'word'
Reason is that expression was meant to describe one word as well
as several words, but considering how hard it would be to parse the
input text into not only single tokens, but also all the different
<i>combinations</i> of tokens, the benefit of selecting multiple
words wouldn't be worth it. It wouldn't be used very often (if ever)
anyway.