Fix API for enriching words to fit the frontend
The data in the frontend has to be structured into an array of arrays.
For readability of the output area, the input text is first divided
into paragraphs, and then into words, which are then enriched through
the API /enrich. If I sent just an array of words, I would have to
assemble them back into the paragraphs after enriching them, which
is more difficult than just changing the API a bit.

If one one paragraph is needed, the data will look like:
[[{},{},{},{},...]] - array of objects in one paragraph